requirements.txt
# Core Web Framework
fastapi~=0.115.8
uvicorn~=0.34.0
starlette~=0.45.3
python-multipart>=0.0.6
pydantic>=2.0.0

# ML & Vector Processing
faiss-cpu~=1.10.0
numpy~=1.26.4
sentence-transformers~=3.4.1

# Image Processing
pillow>=9.0.0

# LLM Integration
torch>=1.8.0
huggingface_hub~=0.24.6
transformers>=4.0.0
llama-cpp-python~=0.3.7

# Template & Utilities
jinja2~=3.1.6
python-dotenv~=1.0.1
posthog~=3.14.1
langdetect~=1.0.9

# Optional: Text utilities (only if needed)
# tokenizers>=0.12.0
# tiktoken>=0.3.0